Solve the token problem
Previous version of code mistakenly encodeURI() the token, at least after Chrome v49.0.2623.87, since the token parser is provided as a herf object, and according to the RFCs, user and password should be transformed.
This version use a regex to abstract the token out when it contains it.
This bugfix only applies to chrome. However, other browsers or user scripts might have the same problem. Developers should realize it and use a more robust way to deal with that.